Core MLOD merged @r157743.
git-svn-id: http://192.168.110.32/svn/pennapps/pff/root/branches/PLF-Release-5.10.0@157768 6963f9ea-81fa-44da-aa10-7ebde50aa541

+++ PFSJava/src/main/java/com/pennanttech/pff/external/service/impl/ExternalInterfaceServiceImpl.java
@@ -1,13 +1,14 @@
 package com.pennanttech.pff.external.service.impl;
 
+import java.util.ArrayList;
 import java.util.List;
 
-import org.apache.commons.collections.CollectionUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 
+import com.pennant.app.constants.ImplementationConstants;
 import com.pennanttech.dataengine.model.DataEngineStatus;
 import com.pennanttech.pennapps.core.resource.Literal;
 import com.pennanttech.pff.external.MandateProcesses;
@@ -32,28 +33,47 @@ public class ExternalInterfaceServiceImpl implements ExternalInterfaceService {
 		if (process_Id != null) {
 			mandateData.setProcess_Id(process_Id);
 			DataEngineStatus status = getMandateProcess().sendReqest(mandateData);
-			if (status.getStatus().equals("F")) {
-				mandateProcessdao.deleteMandateRequests(selectedMandateIds);
-				mandateProcessdao.deleteMandateStatus(selectedMandateIds);
+
+			if (status != null) {
+				if (status.getStatus().equals("F")) {
+					mandateProcessdao.deleteMandateRequests(selectedMandateIds);
+					mandateProcessdao.deleteMandateStatus(selectedMandateIds);
+				}
 			}
 		}
 
 		logger.debug(Literal.LEAVING);
 	}
-
+	
 	@Override
 	public synchronized void processAutoMandateRequest() {
-		logger.debug(Literal.ENTERING);
+		logger.info(Literal.ENTERING);
+		
+		List<Long> mandates = new ArrayList<>();
 		List<String> entityCodes = mandateProcessdao.getEntityCodes();
+		
 		for (String entityCode : entityCodes) {
-			List<Long> mandateList = mandateProcessdao.getMandateList(entityCode);
-			if (CollectionUtils.isNotEmpty(mandateList)) {
-				MandateData mandateData = new MandateData();
-				mandateData.setMandateIdList(mandateList);
-				mandateData.setEntity(entityCode);
-				processMandateRequest(mandateData);
+			if (ImplementationConstants.MANDATE_PTNRBNK_IN_DWNLD) {
+				List<String> partnerBankCodes = mandateProcessdao.getPartnerBankCodeByEntity(entityCode);
+				for (String partnerBankCode : partnerBankCodes) {
+					mandates = mandateProcessdao.getMandateList(entityCode, partnerBankCode);
+					doMandateProcess(entityCode, mandates);
+				}
+			} else {
+				mandates = mandateProcessdao.getMandateList(entityCode);
+				doMandateProcess(entityCode, mandates);
 			}
 		}
+		
+		logger.debug(Literal.LEAVING);
+	}
+
+	private void doMandateProcess(String entityCode, List<Long> mandates) {
+		logger.debug(Literal.ENTERING);
+		MandateData md = new MandateData();
+		md.setEntity(entityCode);
+		md.setMandateIdList(mandates);
+		processMandateRequest(md);
 		logger.debug(Literal.LEAVING);
 	}
